# Optimise dao_get_last_template_usage query